CELE-32 fix: Fix grouping highlighting
