+++ applications/visualizer/frontend/src/components/viewers/TwoD/TwoDViewer.tsx
@@ -5,7 +5,7 @@ import dagre from "cytoscape-dagre";
 import {useSelectedWorkspace} from "../../../hooks/useSelectedWorkspace";
 import {type Connection, ConnectivityService} from "../../../rest";
 import {GRAPH_STYLES} from "../../../theme/twoDStyles";
-import {applyLayout, updateHighlighted} from "../../../helpers/twoD/twoDHelpers";
+import {applyLayout} from "../../../helpers/twoD/twoDHelpers";
 import {
     CHEMICAL_THRESHOLD,
     ELECTRICAL_THRESHOLD,
@@ -18,7 +18,7 @@ import TwoDLegend from "./TwoDLegend";
 import {Box} from "@mui/material";
 import {ColoringOptions, getColor} from "../../../helpers/twoD/coloringHelper";
 import ContextMenu from "./ContextMenu";
-import {computeGraphDifferences} from "../../../helpers/twoD/graphRendering.ts";
+import {computeGraphDifferences, updateHighlighted} from "../../../helpers/twoD/graphRendering.ts";
 
 cytoscape.use(fcose);
 cytoscape.use(dagre);
