+++ applications/visualizer/frontend/src/helpers/twoD/graphRendering.ts
@@ -9,6 +9,7 @@ import {
 } from './twoDHelpers';
 import {NeuronGroup, Workspace} from "../../models";
 import {Connection} from "../../rest";
+import {LegendType} from "../../settings/twoDSettings.tsx";
 
 
 export const computeGraphDifferences = (
@@ -318,4 +319,72 @@ const shouldRemoveEdge = (
     // This approach assumes that expectedNodes has already been processed to reflect all the necessary rules
     // (splitting, joining, and active neuron considerations).
     return !expectedNodes.has(pre) || !expectedNodes.has(post);
+};
+
+
+export const updateHighlighted = (cy, inputIds, selectedIds, legendHighlights) => {
+  // Remove all highlights and return if nothing is selected and no legend item activated.
+  cy.elements().removeClass("faded");
+  if (selectedIds.length === 0 && legendHighlights.size === 0) {
+    return;
+  }
+
+  // Use selected nodes as source if present, otherwise use input nodes.
+  const sourceIds = selectedIds.length ? selectedIds : inputIds;
+  let sourceNodes = cy.collection();
+
+  sourceIds.forEach((id) => {
+    const node = cy.getElementById(id);
+
+    if (node.isParent()) {
+      sourceNodes = sourceNodes.union(node.children());
+    } else {
+      sourceNodes = sourceNodes.union(node);
+    }
+  });
+
+  // Filter network by edges, as set by legend.
+  let edgeSel = "edge";
+  legendHighlights.forEach((highlight, type) => {
+    if (type === LegendType.Connection) {
+      edgeSel += `[type="${highlight}"]`;
+    }else if(type == LegendType.Annotation) {
+      edgeSel += '.' + highlight;
+    }
+  });
+
+  let connectedNodes = sourceNodes.neighborhood(edgeSel).connectedNodes();
+
+  // Filter network by nodes, as set by legend.
+  legendHighlights.forEach((highlight, type) => {
+    if (type === LegendType.Node) {
+      connectedNodes = connectedNodes.filter("[?" + highlight + "]");
+    }
+  });
+
+  // Filter to the neighborhood of the selected nodes.
+  if (selectedIds.length > 0) {
+    let allowedNodes = cy.collection();
+
+    for (let i = 0; i < sourceNodes.length; i++) {
+      const sourceNode = sourceNodes[i];
+      const nodes = sourceNode.neighborhood(edgeSel).connectedNodes();
+
+      if (i === 0) {
+        allowedNodes = allowedNodes.union(nodes);
+      } else {
+        allowedNodes = allowedNodes.intersection(nodes);
+      }
+    }
+    connectedNodes = connectedNodes.intersection(allowedNodes);
+  }
+
+  // Fade out any nodes and edges that were filtered out.
+  let highlightedNodes = sourceNodes.union(connectedNodes);
+  highlightedNodes = highlightedNodes.union(highlightedNodes.parents());
+
+  let highlightedEdges = highlightedNodes.edgesWith(highlightedNodes);
+  highlightedEdges = highlightedEdges.filter(edgeSel);
+
+  cy.elements().not(highlightedNodes).not(highlightedEdges).addClass("faded");
 };
